Make `contacts` optional
I still think it was a good idea to require that each resource has a human that
we can contact about each resource, but nobody else seems to care.

People either ignore it or fill in a generic email account.

I don't have time to track down humans to be responsible for their community
resources, so it's optional now.