Resolve locationSets when reading the files
Because location-conflation has gotten a lot faster in v0.7.0, we can now resolve
all the locationSets into GeoJSON as we read them in reasonable time.

This lets us catch situations where a locationSet resolves to an empty geometry
 e.g. (`include: ['us'], exclude: ['us'])
to prevent mistakes like that from being added to the index in the first place.