Queen Mary uni listing london geometry
Swap the geometry for Queen Mary YouthMappers (Queen Mary University of London group) listing, to use london.geojson the same as the "main" OSMLondon group. This has the effect of demoting it in the listings slightly. Previously this had its own point centred geometry which overlapped most of London resulting a top listing above OSMLondon.