Remove link to global OSM Telegram due to problematic moderation
A mod used "woke" as a slur, and when I and others called it out, our
comments were deleted. I was later muted for 24 hr after posting pride
logos, which were labeled “political” and “off-topic.” Using “woke” as
an insult is often a way to dismiss calls for inclusion and equity, and
silencing those who push back sends the wrong message. Not a space we
should endorse

sources:

https://en.osm.town/@amapanda/114776970953826397
https://en.osm.town/@thibaultmol/114776947333687814