Let's give bun a try
Bun is a new JavaScript runtime with everything built in.
It seems like it could be a lot easier than using a bunch of tools.
It also is a bit faster and has some improvements to security.
This replaces node, c8, esbuild, taze, maybe some other things.

Bun has a builtin http server, so we can use it for local testing.
Just `bun start` or `bun run start` to test location-conflation on localhost.

This commit also moves the locations of of the built files to
`dist/js/*` and `dist/json/*` and stops checking them into git.

Checking generated files into git is problematic, as they change a lot.
These files should really be fetched from the CDN or published package.

The downside is that our `/docs/index.html` preview map will need to
fetch only published data from the CDN, and not recent data from GitHub.
(But you can run this locally now for testing, if needed)